Migrated tokenToResponse to marshal.go
